requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2.3
djangorestframework==3.16.0
exceptiongroup==1.3.0
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.1
pytest-django==4.11.1
requests==2.32.4
soupsieve==2.7
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.14.0
urllib3==2.5.0
